Make exact matches win for path lookup
This fixes a bug where when a wildcard path is specified *after* an
exact path the wildcard would win.  The reason is that for/in loops go
in the order that the keys were defined. Deleting the key and readding
it shows the error. The fix is simply to break out the loop once we find
an exact match. Fixes #189