Implement SystemLoader constructor
This adds a SystemLoader constructor function used internally to inherit
from Loader. `System` then becomes an instance of SystemLoader and
`SystemLoader` is set as `System.constructor`. This way downstream code
can do something like:

```js
class MyLoader extends System.constructor {

  fetch() {

  }

}
```

And doing so will have access to `super` and the like.

I also added in `esnext` to build this and use es6 classes. We talked
about doing it this way and it is *much* more elegant in my opinion
(just look at the minimal amount of code changes).

Let me know if this adds too much code or you don't like it for some
reason and I'll implement this as regular old constructor functions, but
I thought this turned out really well. Fixes #199