Previously, Traceur would return an array of errors when several were detected during compilation. The module loader would just report the first. In 0.9.3 though, Tracuer was updated to instead throw a single "MultipleErrors" type that contained all the errors inside it.
The module loader currently just reports "undefined" as the error, as
there is no array with elements, making for a rather opaque error when
a compilation error occurs.

This modifies it to still throw the first element when dealing with an
array for backwards compatibility, but will otherwise throw the full
exception.